fix: ensure docker images are present
