+++ pkg/docker/fs.go
@@ -1,0 +1,20 @@
+package docker
+
+import (
+	"context"
+	"fmt"
+	"strings"
+
+	"github.com/owenrumney/lazytrivy/pkg/logger"
+	"github.com/owenrumney/lazytrivy/pkg/output"
+)
+
+func (c *Client) ScanFilesystem(ctx context.Context, path string, requiredChecks []string, progress Progress) (*output.Report, error) {
+	logger.Debugf("Scanning filesystem %s", path)
+	checks := strings.Join(requiredChecks, ",")
+
+	progress.UpdateStatus(fmt.Sprintf("Scanning filesystem %s...", path))
+	command := []string{"fs", "--quiet", "--security-checks", checks, "-f=json", "/target"}
+
+	return c.scan(ctx, command, path, []string{}, progress, "lazytrivy:1.0.0", fmt.Sprintf("%s:/target", path))
+}
